# 01-LecturePack/WeatherForecastTestingSolution/FunctionalTests/OpenAPIs/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Weather Forecast Testing 12:45:12 PM",
    "version": "v1"
  },
  "paths": {
    "/api/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Provides a randomly generated set of weather forecasts",
        "description": "Sample request:\r\n\r\nGET /weatherforecast",
        "operationId": "GetAllWeatherForecasts",
        "responses": {
          "200": {
            "description": "Indicates the request was successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecastResult"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Creates a new weather forecast entry and adds it to the list of weather forecasts",
        "operationId": "CreateWeatherForecast",
        "requestBody": {
          "description": "The weather forecast to ad",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastCreate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastCreate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Indicates the weather forecast was added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherForecastResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/WeatherForecast/{id}": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Retrieves the weather forecast by id",
        "operationId": "GetWeatherForecastById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Id of the weather forecast",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Indicates the request was successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherForecast"
                }
              }
            }
          },
          "404": {
            "description": "Indicates the weather forecast was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "PatchWeatherForecastRouteName",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastUpdate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Updates an existing weather forecast with the full payload (idempotent update)\r\nor creates a new weather forecast with a client-provided ID",
        "description": "PUT is used for two scenarios:\r\n1. Full update - Replaces entire resource, requires all fields\r\n2. Create with client-provided ID - Client specifies the resource identifier\r\n\r\nSample request:\r\n\r\nPUT /api/weatherforecast/123e4567-e89b-12d3-a456-426614174000\r\n{\r\n    \"date\": \"2026-01-20T12:00:00Z\",\r\n    \"temperatureC\": 25,\r\n    \"summary\": \"Warm\"\r\n}",
        "operationId": "PutWeatherForecast",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the weather forecast to update or create",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The complete weather forecast data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastCreate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastCreate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherForecastCreate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Indicates the weather forecast was updated successfully"
          },
          "201": {
            "description": "Indicates the weather forecast was created successfully with client-provided ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherForecastResult"
                }
              }
            }
          },
          "400": {
            "description": "Indicates invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Deletes a weather forecast by ID",
        "description": "Sample request:\r\n\r\nDELETE /api/weatherforecast/123e4567-e89b-12d3-a456-426614174000",
        "operationId": "DeleteWeatherForecast",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the weather forecast to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Indicates the weather forecast was deleted successfully"
          },
          "404": {
            "description": "Indicates the weather forecast was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errorNumber": {
            "type": "integer",
            "description": "The error number to be used programmatically",
            "format": "int32"
          },
          "errorMessage": {
            "type": "string",
            "description": "A human readable description of the error for developer to read",
            "nullable": true
          },
          "propertyName": {
            "type": "string",
            "description": "The name of the property involved in the air",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Describe the error response"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "The date of the weather forecast",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "description": "The temperature in Celsius",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "description": "The temperature in Fahrenheit",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "description": "The summary of the weather forecast",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Describes the weather forecast information"
      },
      "WeatherForecastCreate": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "The date of the weather forecast",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "description": "The temperature in Celsius",
            "format": "int32"
          },
          "summary": {
            "maxLength": 60,
            "minLength": 1,
            "type": "string",
            "description": "Summary information about the weather forecast",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Describes the weather forecast information for creation"
      },
      "WeatherForecastResult": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "The date of the weather forecast",
            "format": "date-time"
          },
          "temperatureC": {
            "type": "integer",
            "description": "The temperature in Celsius",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "description": "The temperature in Fahrenheit",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "description": "The summary of the weather forecast",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WeatherForecastUpdate": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "The date of the weather forecast",
            "format": "date-time",
            "nullable": true
          },
          "temperatureC": {
            "type": "integer",
            "description": "The temperature in Celsius",
            "format": "int32",
            "nullable": true
          },
          "summary": {
            "maxLength": 60,
            "minLength": 1,
            "type": "string",
            "description": "Summary information about the weather forecast",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Describes the weather forecast information for update"
      }
    }
  },
  "tags": [
    {
      "name": "WeatherForecast"
    }
  ]
}